test et validation des fonctions
